Rules markdown refactor
We use native fetch to get and cache the rules markdown. Removed the ugly <sup> tag from documentation text. All rules markdown logic moved to his own module.
