+++ package.json
@@ -35,8 +35,6 @@
     "atom-linter": "10.0.0",
     "atom-package-deps": "5.0.0",
     "pluralize": "7.0.0",
-    "request": "2.88.0",
-    "request-promise": "4.2.4",
     "semver": "5.6.0"
   },
   "devDependencies": {
